Move BillingProfile into billingprofile package
