+++ accounting-domain/src/main/java/onlydust/com/marketplace/accounting/domain/model/billingprofile/BillingProfile.java
@@ -1,14 +1,32 @@
-package onlydust.com.marketplace.accounting.domain.model;
+package onlydust.com.marketplace.accounting.domain.model.billingprofile;
 
 import lombok.EqualsAndHashCode;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.NonNull;
+import lombok.experimental.Accessors;
 import lombok.experimental.SuperBuilder;
+import onlydust.com.marketplace.accounting.domain.model.UserId;
 import onlydust.com.marketplace.kernel.model.UuidWrapper;
 
 import java.util.UUID;
 
-public class BillingProfile {
+@Getter
+@Accessors(fluent = true)
+public abstract class BillingProfile {
+
+    @NonNull
+    protected final Id id;
+    @NonNull
+    protected final String name;
+
+    protected BillingProfile(@NonNull String name) {
+        this.id = Id.random();
+        this.name = name;
+    }
+
+    public abstract Type type();
+
     @NoArgsConstructor(staticName = "random")
     @EqualsAndHashCode(callSuper = true)
     @SuperBuilder
@@ -24,5 +42,11 @@ public class BillingProfile {
 
     public enum Type {
         INDIVIDUAL, COMPANY
+    }
+
+    public record User(@NonNull UserId id, @NonNull Role role) {
+        public enum Role {
+            ADMIN, MEMBER
+        }
     }
 }
