+++ accounting-domain/src/main/java/onlydust/com/marketplace/accounting/domain/port/in/BillingProfileFacadePort.java
@@ -1,10 +1,10 @@
 package onlydust.com.marketplace.accounting.domain.port.in;
 
 import lombok.NonNull;
-import onlydust.com.marketplace.accounting.domain.model.BillingProfile;
 import onlydust.com.marketplace.accounting.domain.model.Invoice;
 import onlydust.com.marketplace.accounting.domain.model.RewardId;
 import onlydust.com.marketplace.accounting.domain.model.UserId;
+import onlydust.com.marketplace.accounting.domain.model.billingprofile.BillingProfile;
 import onlydust.com.marketplace.kernel.pagination.Page;
 
 import java.io.InputStream;
