+++ infrastructure/accounting-adapter/src/test/java/onlydust/com/marketplace/api/infrastructure/accounting/AccountingObserverAdapterTest.java
@@ -1,8 +1,8 @@
 package onlydust.com.marketplace.api.infrastructure.accounting;
 
-import onlydust.com.marketplace.accounting.domain.model.BillingProfile;
 import onlydust.com.marketplace.accounting.domain.model.RewardId;
 import onlydust.com.marketplace.accounting.domain.model.RewardStatus;
+import onlydust.com.marketplace.accounting.domain.model.billingprofile.BillingProfile;
 import onlydust.com.marketplace.accounting.domain.port.in.RewardStatusFacadePort;
 import onlydust.com.marketplace.accounting.domain.port.out.RewardStatusStorage;
 import onlydust.com.marketplace.api.domain.model.BillingProfileType;
